feat: loadLikeCount api 구현
